add the base of  createNewRequest, GetAllRequets endpoints and need some edits, new thigns to discuss, (created DTO and converter for request model\!)

+++ Source/src/main/java/org/bringme/repository/RequestRepository.java
@@ -1,5 +1,6 @@
 package org.bringme.repository;
 
+import org.bringme.dto.RequestDTO;
 import org.bringme.model.Request;
 
 import java.util.List;
@@ -7,7 +8,6 @@ import java.util.Optional;
 
 public interface RequestRepository {
     List<Request> getAll();
-    int saveRequest(Request request);
+    Long saveRequest(Request request);
     Optional<Request> getRequestById(Long id);
-    // Search for requests
 }
